Basic styling and API fetch functionality
Finished basic setup of weather app, with ability to search cities. data
for weather in that area is returned via the API to the console. Ran
in to problems with form submission, tried a ton of different fixes only
to find that I had forgotten to preventDefault() on submission. Doh.
Tomorrow I need to go through and look at the data being returned, and
hopefully find a nice way to present it to the user.